+++ Assets/Scripts/Vanilla/GameContent/Artifacts/Chapter2/SweetSleepPillow.cs
@@ -5,6 +5,7 @@ using MVZ2Logic.Entities;
 using MVZ2Logic.Artifacts;
 using MVZ2Logic.Definitions;
 using MVZ2Logic.Entities;
+using MVZ2Logic.Level;
 using PVZEngine.Entities;
 
 namespace MVZ2.GameContent.Artifacts
@@ -18,11 +19,18 @@ namespace MVZ2.GameContent.Artifacts
         public override void PostUpdate(Artifact artifact)
         {
             base.PostUpdate(artifact);
-            artifact.SetGlowing(true);
-            var level = artifact.Level;
-            foreach (var contraption in level.FindEntities(e => e.Type == EntityTypes.PLANT && e.IsFriendlyEntity()))
+            if (artifact.Level.IsAllEnemiesCleared() || artifact.Level.IsCleared)
             {
-                contraption.HealEffects(0.33333333f, contraption);
+                artifact.SetGlowing(false);
+            }
+            else
+            {
+                artifact.SetGlowing(true);
+                var level = artifact.Level;
+                foreach (var contraption in level.FindEntities(e => e.Type == EntityTypes.PLANT && e.IsFriendlyEntity()))
+                {
+                    contraption.HealEffects(0.33333333f, contraption);
+                }
             }
         }
     }