+++ Assets/Scripts/Vanilla/GameContent/Artifacts/Chapter6/CenserOfForgotten.cs
@@ -29,6 +29,9 @@ namespace MVZ2.GameContent.Artifacts
         public override void PostUpdate(Artifact artifact)
         {
             base.PostUpdate(artifact);
+            if (artifact.Level.IsAllEnemiesCleared() || artifact.Level.IsCleared)
+                return;
+
             var secondTimer = artifact.GetSecondTimer();
             if (secondTimer == null)
             {
@@ -64,6 +67,6 @@ namespace MVZ2.GameContent.Artifacts
             }
         }
         public const int MAX_COOLDOWN_SECONDS = 60;
-        public const float STUN_SECONDS = 5;
+        public const float STUN_SECONDS = 10;
     }
 }