+++ Assets/Scripts/Vanilla/GameContent/Stages/Behaviours/FinalClear/RedDragonStageBehaviour.cs
@@ -65,6 +65,7 @@ namespace MVZ2.GameContent.Stages
         {
             base.AfterBossWaveUpdate(level);
             ClearEnemies(level);
+            EnterLevelEnemiesClearedState(level);
 
             if (!level.IsFirstAdventure())
                 return;