+++ Assets/Scripts/Vanilla/GameContent/Stages/Behaviours/IZombie/IZombieEndless2Behaviour.cs
@@ -4,11 +4,8 @@ using System.Collections.Generic;
 using MVZ2.GameContent.Enemies;
 using MVZ2.GameContent.Seeds;
 using MVZ2Logic.Blueprints;
-using MVZ2Logic.Level;
-using MVZ2Logic.Blueprints;
 using PVZEngine;
 using PVZEngine.Level;
-using PVZEngine.Definitions;
 
 namespace MVZ2.GameContent.Stages
 {
@@ -19,18 +16,18 @@ namespace MVZ2.GameContent.Stages
         }
         protected override IEnumerable<IZELayoutItem> GetNormalLayouts()
         {
-            // Aï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Composite, 1.2f); // ï¿½Ûºï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Control, 1.2f); // ï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Spectral, 0.8f); // ï¿½ï¿½ï¿½é£¨ï¿½ï¿½1ï¿½ï¿½Ò©Í°ï¿½ï¿½
+            // AÀàÕó
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Composite, 1.2f); // ×ÛºÏ
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Control, 1.2f); // ¿ØÖÆ
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Spectral, 1.0f); // ¹âÁé£¨ÓÐ1»ðÒ©Í°£©
 
-            // Bï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Spikes, 0.4f); // ï¿½ï¿½Ð±
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Dispensers, 0.4f); // ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½1ï¿½ï¿½Ò©Í°ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Defense, 0.4f); // ï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Fire, 0.4f); // ï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Instakill, 0.2f); // ï¿½ï¿½ï¿½ï¿½
-            yield return new IZELayoutItem(VanillaIZombieLayoutID.izeAwards, 0.2f); // ï¿½Ö¸ï¿½
+            // BÀàÕó
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Spikes, 0.4f); // ÇãÐ±
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Dispensers, 0.4f); // ·¢ÉäÆ÷£¨ÓÐ1»ðÒ©Í°£©
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Impale, 0.4f); // ´©´Ì
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Fire, 0.4f); // »ðÑæ
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.ize2Instakill, 0.2f); // ¼´ËÀ
+            yield return new IZELayoutItem(VanillaIZombieLayoutID.izeAwards, 0.2f); // »Ö¸´
         }
         protected override IEnumerable<IZELayoutItem> GetAwardLayouts()
         {